Move config to file ...
... and encapsulated the angle to servo position calculations as well as
limit checks for each servo. Configuration for these settings is now in
the evoarm.yaml config file. In theory this allows for use of externally
geared servos - the Dynamixels are fast but not very powerful or
accurate, so a possible plan is to reduce the main servos by 2:1 to use
their full range of motion with twice the positional accuracy and power.